app/src/main/java/com/crushcoder/calculator/DisplayFragment.java
package com.crushcoder.calculator;


import android.app.Fragment;
import android.os.Bundle;
import android.support.annotation.Nullable;
import android.view.LayoutInflater;
import android.view.View;
import android.view.ViewGroup;

public class DisplayFragment extends Fragment {
    public DisplayFragment() {
    }

    @Nullable
    @Override
    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
        View view = inflater.inflate(R.layout.fragment_display, container, false);
        return view;
    }

    public static DisplayFragment newInstance() {
        return new DisplayFragment();
    }
}
